stylegan3/requirements.txt
clean_fid
click
glfw==2.2.0
imageio
imgui==1.3.0
lmdb
matplotlib
numpy
opencv_python
Pillow
psutil
PyMatting
PyOpenGL==3.1.5
pyspng
Requests
scipy
torch==1.9.1
torchvision
tqdm
